backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
# Use Motor 2.x which is compatible with pymongo 3.12.x
motor==2.5.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sendgrid==6.11.0
python-dotenv==1.0.0
pymongo==3.12.3
email-validator==1.3.1
